+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.3.63](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.62..v0.3.63) - March 6, 2023
+
+### Fixed
+
+- Make all taxonomies under "Fächer im Aufbau" subjects
+
+## [v0.3.62](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.61..v0.3.62) - March 2, 2023
+
+### Fixed
+
+- Make all taxonomies under "Mathematik > Partner" as subjects
+
+## [v0.3.61](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.60..v0.3.61) - March 2, 2023
+
+### Added
+
+- Add serlo.org/268835 "Chancenwerk" as subject
+
+## [v0.3.60](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.59..v0.3.60) - February 2, 2023
+
+### Fixed
+
+- thread: fix name of ThreadEditMutation
+
 ## [v0.3.59](https://github.com/serlo/serlo.org-database-layer/compare/v0.3.58..v0.3.59) - January 26, 2023
 
 ### Added